Changed structure for bot, added some docstrings.
